// src/generated/roles.ts
// @generated by `npm run protocol:generate`; do not edit.
export type Role = { id: string, revision: number, name: string, description: string, images: Array<string>, createdAt: string, updatedAt: string, };

export type RoleInlineImage = { name: string, base64: string, };

export type RoleSaveRequest = { id: string | null, revision: number | null, name: string, description: string, images: Array<string>, inlineImages: Array<RoleInlineImage>, };